fetch data to Geocode and pass to Darksky API
