+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import logo from './logo.svg';
 import './App.css';
-import Weather from './Weather';
+import Weather from './components/Weather';
 class App extends Component {
   render() {
     return (
